fancy plot for talks

+++ Cluster_Markers_Manual_ID.R
@@ -25,6 +25,60 @@ Cluster_guesses <- paste(0:(length(assign_general)-1),".",rep("Unknown", length(
 Cluster_guesses[assign_general > 0.5] <- names(assign_general)[assign_general>0.5]
 Cluster_guesses[assign_specific > 0.5] <- names(assign_specific)[assign_specific>0.5]
 
+## Figures for Talk ##
+agg_coord_by_cluster <- function(coords, clusters) {
+	x <- split(seq(nrow(coords)), clusters)
+	result <- sapply(x, function(a) apply(coords[a, ],2,median))
+	return(result)
+}
+
+
+
+lab_pos <- agg_coord_by_cluster(int.obj@reductions$tsne@cell.embeddings, int.obj@meta.data$Fine_clusters)
+lab_id <- colnames(lab_pos)
+clust_autoanno <- sapply(strsplit(Cluster_guesses, "\\."), function(a){a[[2]]})
+clust_autoanno[clust_autoanno == "UnidentifiedHep"] <- "Hepatocyte";
+clust_autoanno[clust_autoanno == "CD3abTcells"] <- "Tcells";
+clust_autoanno[clust_autoanno == "PericentralHep"] <- "Hepatocyte";
+clust_autoanno[clust_autoanno == "Unknown"] <- "Hepatocyte";
+clust_autoanno[clust_autoanno == "AntibodysecretingBcells"] <- "Bcells";
+clust_autoanno[clust_autoanno == "CentralvenousLSECs"] <- "cvLSEC";
+clust_autoanno[clust_autoanno == "InflamatoryMacrophages"] <- "inf-Macrophage";
+clust_autoanno[clust_autoanno == "Non-inflammatoryMacrophages"] <- "non-Macrophage";
+
+
+Colour_scheme <- rbind(
+	c("cvLSECs", "goldenrod1"),
+	c("LSECs", "violet"),
+	c("Bcells", "darkorchid"),
+	c("Tcells", "forestgreen"),
+	c("Macrophage", "cornflowerblue"),
+	c("InflamatoryMacrophages", "navy"),
+	c("Non-inflamatoryMacrophages", "dodgerblue"),
+	c("Erythoidcells", "darkred"),
+	c("Hepatocytes", "firebrick2"))
+Colour_scheme <- Colour_scheme[order(Colour_scheme[,1]),]
+
+int.obj@meta.data$autoanno_c <- clust_autoanno[int.obj@meta.data$Fine_clusters]
+
+require("ggplot2")
+png("General_autoanno_label_harmony_tsne.png", width=7.5, height=6, units="in", res=300)
+DimPlot(int.obj, reduction = "tsne", group.by = "autoanno_c", pt.size = .1)+scale_color_manual(values=Colour_scheme[,2])+annotate("text", x=lab_pos[1,], y=lab_pos[2,], label=lab_id, colour="grey35")
+dev.off();
+
+
+metadata <- read.table("Metadata20LiverMapPlusParams.csv", sep=",", header=T)
+
+pats <- levels(factor(int.obj@meta.data$donor))
+sex <- as.character(metadata[match(pats, metadata$Name),"sex"])
+sex[!sex %in% c("M", "F")] <- "Unk"
+int.obj@meta.data$sex <- sex[factor(int.obj@meta.data$donor)]
+
+png("General_donor_label_harmony_tsne.png", width=6.5, height=6, units="in", res=300)
+DimPlot(int.obj, reduction = "tsne", group.by = "donor", pt.size = .1)
+dev.off();
+
+DimPlot(int.obj, reduction = "tsne", group.by = "sex", pt.size = .1)+scale_color_manual(values=c("salmon","cornflowerblue", "grey50"))
 # Output:
 
 # Auto-anno table:
@@ -56,52 +110,46 @@ my_rowSums <- function(x) {
 }
 
 ## Significant with MAST
-set.seed(8817)
+if (!file.exists("harmony_MASTmodel.rds")) {
+	set.seed(8817)
 
-require("MAST")
+	require("MAST")
 
-sca <- FromMatrix(as.matrix(int.obj@assays$RNA@data)) ##### <---- not sparse this is very Mem heavy!!
+	sca <- FromMatrix(as.matrix(int.obj@assays$RNA@data)) ##### <---- not sparse this is very Mem heavy!!
 
-ngenes <- colSums(int.obj@assays$RNA@counts > 0)
-ncounts <- colSums(int.obj@assays$RNA@counts)
-donor <- int.obj@meta.data$donor
-cluster <- int.obj@meta.data$Fine_clusters
+	ngenes <- colSums(int.obj@assays$RNA@counts > 0)
+	ncounts <- colSums(int.obj@assays$RNA@counts)
+	donor <- int.obj@meta.data$donor
+	cluster <- int.obj@meta.data$Fine_clusters
 
-# This is slow with so many cells.... Perhaps move to H4h cluster.
-mod <- zlm(~cluster + donor + ngenes + ncounts, sca)
+	# This is slow with so many cells.... Perhaps move to H4h cluster.
+	mod <- zlm(~cluster + donor + ngenes + ncounts, sca)
+
+	saveRDS(mod, "harmony_MASTmodel.rds")
+}
 
 
-saveRDS(mod, "harmony_MASTmodel.rds")
+#res1 <- summary(mod, doLRT="cluster1")
+#res <- res1$datatable[res1$datatable$contrast=="cluster1" & res1$datatable$component %in% c("C","D"),]
+#res$fdr <- p.adjust(unlist(res[,4]), method="fdr")
 
-set.seed(3891)
+#con <- res[res$component =="C",]
+#dis <- res[res$component =="D",]
 
-res1 <- summary(mod, doLRT="cluster1")
-res <- res1$datatable[res1$datatable$contrast=="cluster1" & res1$datatable$component %in% c("C","D"),]
-res$fdr <- p.adjust(unlist(res[,4]), method="fdr")
-
-con <- res[res$component =="C",]
-dis <- res[res$component =="D",]
-
-sig <- (con$fdr < 0.01 | dis$fdr < 0.01) & sign(con$coef) == sign(dis$coef)
-out <- cbind(con[sig,], dis[sig,])
-out$overall <- out[,7]+out[,16]
-out <- out[order(out$overall, decreasing=T),]
-write.table(out, file=paste(this_contrast,"_MAST_output.txt", sep=""), col.names=T, row.names=F)
+#sig <- (con$fdr < 0.01 | dis$fdr < 0.01) & sign(con$coef) == sign(dis$coef)
+#out <- cbind(con[sig,], dis[sig,])
+#out$overall <- out[,7]+out[,16]
+#out <- out[order(out$overall, decreasing=T),]
+#write.table(out, file=paste(this_contrast,"_MAST_output.txt", sep=""), col.names=T, row.names=F)
  
 
-
-res2 <- summary(mod, doLRT="cluster2")
-res3 <- summary(mod, doLRT="cluster3")
-res4 <- summary(mod, doLRT="cluster4")
-res5 <- summary(mod, doLRT="cluster5")
-res6 <- summary(mod, doLRT="cluster6")
 
 ## Pseudobulks ##
 make_pseudobulk_table <- function(mat, clusters, indis) {
 	c <- split(seq(ncol(mat)), clusters);
 	donor_expr <- function(MAT, ds) {
 		d <- split(seq(ncol(MAT)), ds);
-		mus <- sapply(d, function(group) my_rowMeans(MAT[,group]))
+		mus <- sapply(d, function(group) my_rowSums(MAT[,group]))
 		return(mus);
 	}
 	pseudobulk_expr <- lapply(c, function(clust) {
@@ -112,11 +160,40 @@ make_pseudobulk_table <- function(mat, clusters, indis) {
 }
 
 make_pseudobulk_v2 <- function(mat, clusters, indis) {
-	tab <- apply(mat, 1, function(x){tapply(x, list(clusters, donors), mean)})
-	rownames(tab) <- paste(rep(levels(as.factor(clusters)), each=length(levels(as.factor(donors)))), 
-			rep(levels(as.factor(donors)), times=length(levels(as.factor(clusters)))), sep="_")
-	return(t(tab))
+	tab <- apply(mat, 1, function(x){tapply(x, list(clusters, indis), sum)})
+	c_lab <- rep(levels(as.factor(clusters)), each=length(levels(as.factor(indis))))
+	i_lab <- rep(levels(as.factor(indis)), times=length(levels(as.factor(clusters))))
+	rownames(tab) <- paste(c_lab, i_lab, sep="_")
+	return(list(bulks=t(tab), groups=c_lab, indis=i_lab));
 }
+
+pseudobulk <- make_pseudobulk_v2(int.obj@assays$RNA@counts,  int.obj@meta.data$Fine_clusters, int.obj@meta.data$donor)
+
+# subset
+keep <- which(!is.na(colSums(pseudobulk$bulks)));
+pseudobulk$bulks <- pseudobulk$bulks[,keep]
+pseudobulk$groups <- pseudobulk$groups[keep]
+pseudobulk$indis <- pseudobulk$indis[keep]
+
+
+# run edger - No longer super biased but I'm still not happy with the results: l2fc are really high significance is really low.....
+cluster = "1"
+coef = paste("groups",cluster,sep="")
+
+require("edgeR")
+edger.obj <- DGEList(counts=pseudobulk$bulks, group=pseudobulk$groups);
+edger.obj <- calcNormFactors(edger.obj)
+design <- model.matrix(~pseudobulk$groups + pseudobulk$indis)
+edger.obj <- estimateDisp(edger.obj, design)
+fit <- glmQLFit(edger.obj, design) #glmQLFit/glmFit
+contrast_vec <- rep(-1, ncol(design));
+this_group = which(grepl(coef, colnames(design)));
+if (length(this_group) > 0) {
+	contrast_vec[this_group[1]] <- 1;
+}
+contrast_vec[1] <- 0
+lrt <- glmQLFTest(fit, contrast=contrast_vec) #glmQLFTest/glmLRT
+topTags(lrt, 100)
 
 
 ## Ranked list using my method ##
